DA: Refactor Folder Structure
