+++ apps/server/src/core/services/user.service.ts
@@ -42,20 +42,26 @@ export class UserService {
   getUserRoles() {}
 
   private getUserFromAuth0() {
-    // this.httpService
-    //   .get('https://dev-oixkoo26y6fn8sbd.uk.auth0.com/userinfo', {
-    //     headers: {
-    //       'content-type': 'application/json',
-    //       Authorization: this.request.headers.authorization,
-    //     },
-    //   })
-    //   .subscribe(
-    //     (res) => {
-    //       console.log(JSON.stringify(res));
-    //     },
-    //     (error) => {
-    //       console.log(error);
-    //     },
-    //   );
+    this.httpService
+      .get('https://dev-oixkoo26y6fn8sbd.uk.auth0.com/userinfo', {
+        headers: {
+          'content-type': 'application/json',
+          Authorization: this.request.headers.authorization,
+        },
+      })
+      .subscribe(
+        (res) => {
+          const responseData = res.data;
+          console.log(JSON.stringify(responseData));
+          // const user: NonDBUser = {
+          //   nickname: responseData.nickname,
+          //   email: responseData.email,
+          //   emailVerified: responseData.emailVerified,
+          // };
+        },
+        (error) => {
+          console.log(error);
+        },
+      );
   }
 }
