+++ packages/database-prisma/src/transactions/consent-request.da.ts
@@ -28,113 +28,117 @@ import {
   DataSchema,
   Optional,
 } from "@consent-as-a-service/domain";
-import { mapConsentRequestToModel } from "../mappers/consent-request.mapper";
+import { mapConsentRequestToModel } from "../internal/mappers/consent-request.mapper";
 import { container } from "tsyringe";
-import { PrismaClientService } from "../internal/prisma-client.service";
-import { TransactionDaInternal } from "../internal/transaction.da.internal";
+import { PrismaClientService } from "../internal/services/prisma-client.service";
+import { TransactionDaInternal } from "../internal/prismada/transaction.da.internal";
 import {
   ConsentRequestDaInternal,
   CreateConsentRequestOptions,
   CreateConsentResultType,
-} from "../internal/consent-request.da.internal";
-import { ConsentDataSchemaDaInternal } from "../internal/consent-data-schema.da.internal";
+} from "../internal/prismada/consent-request.da.internal";
+import { ConsentDataSchemaDaInternal } from "../internal/prismada/consent-data-schema.da.internal";
 
-async function getServices() {
-  const prismaClientService = container.resolve(PrismaClientService);
-  const txnDa = container.resolve(TransactionDaInternal);
-  const internalDa = container.resolve(ConsentRequestDaInternal);
-  const schemaDa = container.resolve(ConsentDataSchemaDaInternal);
-  return { prismaClientService, txnDa, internalDa, schemaDa };
-}
-
-/**
- * High Level abstraction for creating consent requests on the server.
- * Will take a Consent Model and will also handle setting up a new txn for this request
- * Consent Requests are **Immutable** and cannot be updated once set.
- * @returns
- */
-export const CreateConsentRequestType = async (
-  options: CreateConsentRequestOptions,
-  dataSchema: Array<DataEntry<any>>
-): Promise<CreateConsentResultType> => {
-  console.log("Creating a new Consent Request Type");
-  const { prismaClientService, txnDa, internalDa, schemaDa } =
-    await getServices();
-  await prismaClientService.connect();
-  const txn = await txnDa.createTxn({
-    txnStatus: "CREATED",
-  });
-  const schemaEntry = await schemaDa.createSchemaEntry(dataSchema);
-  const schema = {
-    id: schemaEntry.typeId,
-    entries: dataSchema,
-  } as DataSchema;
-
-  const dbR: ConsentRequests = await internalDa.createConsentRequestType(
-    txn.txnId,
-    {
-      title: options.title,
-      description: options.description,
-      schema: schema,
-      callbackUrl: options.callbackUrl,
-    }
-  );
-  await prismaClientService.disconnect();
-  return {
-    txnId: txn.txnId,
-    consentRequestId: dbR.consentRequestId,
-    schemaId: schemaEntry.typeId,
-  } as CreateConsentResultType;
-};
-
-/**
- * Reads out the txn with a deep link to a TxnModel
- * @param consentRequestId
- * @constructor
- */
-export const ReadConsentRequest = async (
-  consentRequestId: NonNullable<string>
-): Promise<Optional<ConsentRequestModel>> => {
-  console.log("Reading a ConsentRequest");
-  const { prismaClientService, txnDa, internalDa, schemaDa } =
-    await getServices();
-  await prismaClientService.connect();
-  const req: Optional<ConsentRequests> = await internalDa.readConsentRequest(
-    consentRequestId
-  );
-  await prismaClientService.disconnect();
-  if (!req.isPresent()) {
-    return Optional.empty();
+export namespace ConsentRequestDA {
+  async function getServices() {
+    const prismaClientService = container.resolve(PrismaClientService);
+    const txnDa = container.resolve(TransactionDaInternal);
+    const internalDa = container.resolve(ConsentRequestDaInternal);
+    const schemaDa = container.resolve(ConsentDataSchemaDaInternal);
+    return { prismaClientService, txnDa, internalDa, schemaDa };
   }
-  //Get the Data
-  const schema = await schemaDa.readSchema(req.get().dataId);
-  const txnModel: Optional<TxnLog> = await txnDa.readTxn(req.get().txnId);
-  //Map to Domain
-  return Optional.of(
-    mapConsentRequestToModel(req.get(), txnModel.get(), schema.get())
-  );
-};
 
-/**
- * Voids the Consent Request (makes permanently unusable) via updating its Txn to VOIDED
- * @param consentRequestId
- * @constructor
- */
-export const VoidConsentRequest = async (
-  consentRequestId: NonNullable<string>
-): Promise<string> => {
-  const { prismaClientService, txnDa, internalDa } = await getServices();
-  await prismaClientService.connect();
-  const consentRequestModel: Optional<ConsentRequests> =
-    await internalDa.readConsentRequest(consentRequestId);
-  await prismaClientService.disconnect();
-  if (consentRequestModel.isPresent()) {
-    const txnId = consentRequestModel.get().txnId;
-    const newTxn = await txnDa.updateTxn(txnId, {
-      txnStatus: "VOIDED",
+  /**
+   * High Level abstraction for creating consent requests on the server.
+   * Will take a Consent Model and will also handle setting up a new txn for this request
+   * Consent Requests are **Immutable** and cannot be updated once set.
+   * @returns
+   */
+  export const CreateConsentRequestType = async (
+    options: CreateConsentRequestOptions,
+    dataSchema: Array<DataEntry<any>>
+  ): Promise<CreateConsentResultType> => {
+    console.log("Creating a new Consent Request Type");
+    const { prismaClientService, txnDa, internalDa, schemaDa } =
+      await getServices();
+    await prismaClientService.connect();
+    const txn = await txnDa.createTxn({
+      txnStatus: "CREATED",
     });
-    return newTxn.txnId;
-  } else {
-    throw new Error("Unable to void request, no matching transaction present");
-  }
-};
+    const schemaEntry = await schemaDa.createSchemaEntry(dataSchema);
+    const schema = {
+      id: schemaEntry.typeId,
+      entries: dataSchema,
+    } as DataSchema;
+
+    const dbR: ConsentRequests = await internalDa.createConsentRequestType(
+      txn.txnId,
+      {
+        title: options.title,
+        description: options.description,
+        schema: schema,
+        callbackUrl: options.callbackUrl,
+      }
+    );
+    await prismaClientService.disconnect();
+    return {
+      txnId: txn.txnId,
+      consentRequestId: dbR.consentRequestId,
+      schemaId: schemaEntry.typeId,
+    } as CreateConsentResultType;
+  };
+
+  /**
+   * Reads out the txn with a deep link to a TxnModel
+   * @param consentRequestId
+   * @constructor
+   */
+  export const ReadConsentRequest = async (
+    consentRequestId: NonNullable<string>
+  ): Promise<Optional<ConsentRequestModel>> => {
+    console.log("Reading a ConsentRequest");
+    const { prismaClientService, txnDa, internalDa, schemaDa } =
+      await getServices();
+    await prismaClientService.connect();
+    const req: Optional<ConsentRequests> = await internalDa.readConsentRequest(
+      consentRequestId
+    );
+    await prismaClientService.disconnect();
+    if (!req.isPresent()) {
+      return Optional.empty();
+    }
+    //Get the Data
+    const schema = await schemaDa.readSchema(req.get().dataId);
+    const txnModel: Optional<TxnLog> = await txnDa.readTxn(req.get().txnId);
+    //Map to Domain
+    return Optional.of(
+      mapConsentRequestToModel(req.get(), txnModel.get(), schema.get())
+    );
+  };
+
+  /**
+   * Voids the Consent Request (makes permanently unusable) via updating its Txn to VOIDED
+   * @param consentRequestId
+   * @constructor
+   */
+  export const VoidConsentRequest = async (
+    consentRequestId: NonNullable<string>
+  ): Promise<string> => {
+    const { prismaClientService, txnDa, internalDa } = await getServices();
+    await prismaClientService.connect();
+    const consentRequestModel: Optional<ConsentRequests> =
+      await internalDa.readConsentRequest(consentRequestId);
+    await prismaClientService.disconnect();
+    if (consentRequestModel.isPresent()) {
+      const txnId = consentRequestModel.get().txnId;
+      const newTxn = await txnDa.updateTxn(txnId, {
+        txnStatus: "VOIDED",
+      });
+      return newTxn.txnId;
+    } else {
+      throw new Error(
+        "Unable to void request, no matching transaction present"
+      );
+    }
+  };
+}
