+++ packages/database-prisma/src/transactions/index.ts
@@ -23,3 +23,4 @@
 
 export * from "./transaction.da";
 export * from "./consent-request.da";
+export * from "./transaction.da";
